[frontend] - refactor/fix: Refactor login page and Fix `Logo` component
