+++ frontend/src/components/Logo/Logo.test.jsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
 
-import Logo from "@/components/Text/Logo/Logo";
+import Logo from "@/components/Logo/Logo";
 
 describe("Logo", () => {
   it("should create", () => {
