+++ frontend/src/pages/Login/Login.test.jsx
@@ -1,5 +1,5 @@
 import { customRender } from "@/utilities/tests/customRender";
-import Login from "@/components/pages/Login/Login";
+import Login from "@/pages/Login/Login";
 
 describe("Login", () => {
   it("should create", () => {
